+++ relay-server/src/service.rs
@@ -8,7 +8,9 @@ use crate::services::buffer::{
     ObservableEnvelopeBuffer, PartitionedEnvelopeBuffer, ProjectKeyPair,
 };
 use crate::services::cogs::{CogsService, CogsServiceRecorder};
-use crate::services::global_config::{GlobalConfigManager, GlobalConfigService};
+use crate::services::global_config::{
+    GlobalConfigHandle, GlobalConfigManager, GlobalConfigService,
+};
 use crate::services::health_check::{HealthCheck, HealthCheckService};
 use crate::services::metrics::RouterService;
 #[cfg(feature = "processing")]
@@ -82,6 +84,7 @@ pub struct Registry {
     #[cfg(feature = "processing")]
     pub objectstore: Option<Addr<Objectstore>>,
     pub upload: Addr<Upload>,
+    pub global_config_handle: GlobalConfigHandle,
 }
 
 /// Constructs a Tokio [`relay_system::Runtime`] configured for running [services](relay_system::Service).
@@ -295,7 +298,7 @@ impl ServiceState {
                     EnvelopeProcessorService::new(
                         processor_pool.clone(),
                         config.clone(),
-                        global_config_handle,
+                        global_config_handle.clone(),
                         project_cache_handle.clone(),
                         cogs,
                         #[cfg(feature = "processing")]
@@ -374,6 +377,7 @@ impl ServiceState {
             #[cfg(feature = "processing")]
             objectstore,
             upload,
+            global_config_handle,
         };
 
         let state = StateInner {
@@ -458,6 +462,10 @@ impl ServiceState {
     pub fn upload(&self) -> &Addr<Upload> {
         &self.inner.registry.upload
     }
+
+    pub fn global_config_handle(&self) -> &GlobalConfigHandle {
+        &self.inner.registry.global_config_handle
+    }
 }
 
 /// Creates Redis clients from the given `configs`.
@@ -469,7 +477,6 @@ impl ServiceState {
 /// is created for each use case.
 #[cfg(feature = "processing")]
 pub fn create_redis_clients(configs: RedisConfigsRef<'_>) -> Result<RedisClients, RedisError> {
-    const CARDINALITY_REDIS_CLIENT: &str = "cardinality";
     const PROJECT_CONFIG_REDIS_CLIENT: &str = "projectconfig";
     const QUOTA_REDIS_CLIENT: &str = "quotas";
     const UNIFIED_REDIS_CLIENT: &str = "unified";
@@ -480,23 +487,19 @@ pub fn create_redis_clients(configs: RedisConfigsRef<'_>) -> Result<RedisClients
 
             Ok(RedisClients {
                 project_configs: client.clone(),
-                cardinality: client.clone(),
                 quotas: client,
             })
         }
         RedisConfigsRef::Individual {
             project_configs,
-            cardinality,
             quotas,
         } => {
             let project_configs =
                 create_async_redis_client(PROJECT_CONFIG_REDIS_CLIENT, &project_configs)?;
-            let cardinality = create_async_redis_client(CARDINALITY_REDIS_CLIENT, &cardinality)?;
             let quotas = create_async_redis_client(QUOTA_REDIS_CLIENT, &quotas)?;
 
             Ok(RedisClients {
                 project_configs,
-                cardinality,
                 quotas,
             })
         }
@@ -525,10 +528,13 @@ async fn initialize_redis_scripts_for_client(
 ) -> Result<(), RedisError> {
     let scripts = RedisScripts::all();
 
-    let clients = [&redis_clients.cardinality, &redis_clients.quotas];
-    for client in clients {
-        initialize_redis_scripts(client, &scripts).await?;
-    }
+    let RedisClients {
+        project_configs,
+        quotas,
+    } = redis_clients;
+
+    initialize_redis_scripts(project_configs, &scripts).await?;
+    initialize_redis_scripts(quotas, &scripts).await?;
 
     Ok(())
 }
@@ -536,7 +542,7 @@ async fn initialize_redis_scripts_for_client(
 #[cfg(feature = "processing")]
 async fn initialize_redis_scripts(
     client: &AsyncRedisClient,
-    scripts: &[&Script; 3],
+    scripts: &[&Script],
 ) -> Result<(), RedisError> {
     let mut connection = client.get_connection().await?;